:recycle: (s3) Move S3 related helpers to a different…
